types: move AliasedType and BuiltinAlias to aliased.rs
